feat: build AST for resources with typescript compiler
